+++ Samples/LedLineIt/Sources/FontMenu.cpp
@@ -34,7 +34,7 @@ void FixupFontMenu (CMenu* fontMenu)
     }
 
     vector<Led_SDK_String> fontNames = sInstalledFonts.GetUsableFontNames ();
-    for (size_t i = 0; i < fontNames.size (); i++) {
+    for (UINT i = 0; i < fontNames.size (); i++) {
         int cmdNum = cmdFontMenuFirst + i;
         fontMenu->AppendMenu (MF_STRING, cmdNum, fontNames[i].c_str ());
     }
